refactor(cypher): split evaluate_scalar_function into per-category modules
The 2,573-line scalar_functions.rs (a single ~2,169-line match over 119
function names) becomes a scalar_functions/ module directory:

- mod.rs   — thin dispatcher delegating to each category in turn, plus the
             shared helper methods (eval_local_temporal, node_arg_index,
             resolve_timeseries_* , extract_float_list).
- shared.rs — free helpers/const/enum (json_to_value, next_random_*,
             coerce_naive_datetime, LocalTemporalKind, SPATIAL_RESOLUTION_HELP).
- string/numeric/temporal/graph/spatial/collection/timeseries/utility.rs —
             one eval_<cat>_fn each, returning Ok(None) when it doesn't own
             the name so the dispatcher falls through to the next.

Match arms are moved verbatim (no behaviour change); the only edits are the
per-category Ok(None) fallback and wrapping explicit `return Ok(x)` as
`return Ok(Some(x))` to fit the Result<Option<Value>, String> shape. Verified
the set of function-name literals is identical before/after. Largest new file
is 424 lines; no test_god_files allowlist change.
